refactor(workflow): moved presentation to separate file Signed-off-by: Stefan Genov <stefan.genov@broadcom.com>
